+++ libs/sabi-core/src/auth/v4/calculator/can_generate_mac.rs
