+++ libs/sabi-core/src/auth/v4/calculator/output_hmac.rs
