+++ libs/sabi-core/src/auth/v4/mod.rs
@@ -1,4 +1,4 @@
+pub mod calculator;
 pub mod canonical;
 pub mod chrono;
-pub mod key;
 pub mod sign;
